Modify structure of repo; add testpypi req files
