+++ src/modi_flows.egg-info/SOURCES.txt
@@ -1,0 +1,26 @@
+.gitignore
+LICENSE
+README.md
+pyproject.toml
+setup.py
+data/input/classlabels.txt
+data/input/JF30_subset/JFT_00020.jpg
+data/input/JF30_subset/JFT_00023.jpg
+data/input/JF30_subset/JFT_00170.jpg
+data/input/JF30_subset/JFT_00173.jpg
+data/input/JF30_subset/JFT_00938.jpg
+data/input/JF30_subset/JFT_00941.jpg
+data/input/JF30_subset/JFT_01250.jpg
+data/input/JF30_subset/JFT_01256.jpg
+data/input/JF30_subset/JFT_01476.jpg
+data/input/JF30_subset/JFT_01479.jpg
+misc/POSTER_MODI.pdf
+src/modi_flows/__init__.py
+src/modi_flows/dynamics.py
+src/modi_flows/initialization.py
+src/modi_flows/main.py
+src/modi_flows.egg-info/PKG-INFO
+src/modi_flows.egg-info/SOURCES.txt
+src/modi_flows.egg-info/dependency_links.txt
+src/modi_flows.egg-info/requires.txt
+src/modi_flows.egg-info/top_level.txt
